printf

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: gyopark <gyopark@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/26 16:52:56 by gyopark           #+#    #+#             */
-/*   Updated: 2022/11/27 17:53:21 by gyopark          ###   ########.fr       */
+/*   Updated: 2022/11/28 20:28:50 by gyopark          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,8 +29,8 @@ ssize_t	ft_print_sx(va_list **ap);
 ssize_t	ft_print_bx(va_list **ap);
 ssize_t	ft_print_p(va_list **ap);
 
-char	*ft_change_hexa(long long vatemp);
-size_t	ft_hexalen(long long num);
-char	*ft_change_hexa_address(long long vatemp);
+char	*ft_change_hexa(unsigned int vatemp);
+size_t	ft_hexalen(size_t num);
+char	*ft_change_hexa_address(size_t vatemp);
 
 #endif
